Extract quantum encryption logic to pkg/quantum package
Refactor quantum-resistant cryptographic operations into modular package:
- pkg/quantum/types.go: Core types and ML-KEM constants
- pkg/quantum/mlkem.go: ML-KEM-1024 key generation
- pkg/quantum/hybrid.go: Hybrid ML-KEM + AES-256-GCM encryption
- pkg/quantum/security.go: Security utilities (SecureZero, SecureCompare)

Maintains 100% functional compatibility and security properties:
- Zero logic changes to quantum encryption algorithms
- Preserves memory security and constant-time operations
- Keeps ML-KEM-1024 + AES-256-GCM hybrid architecture
- Ensures proper key zeroing and side-channel resistance
